Minor fixes for new board Added ability to read volatile memory Added ability to toggle SX switch power for debugging Fixed and improved current based PWM limiting

+++ RemoteStation/PermanentStorage.h
@@ -62,7 +62,7 @@ typedef struct PermanentVariables
 
   signed char tsOffset;
   byte tsGain;
-  int wdCalibMin, 
+  short wdCalibMin, 
     wdCalibMax;
 
   unsigned short chargeVoltage_mV;
